feat: update savingsplan module
