+++ Features/Sources/SavingsPlanModule/Network/SavingsPlanService.swift
@@ -8,6 +8,7 @@
 import Foundation
 import NetworkKit
 import CoreModule
+import Models
 
 struct SavingsPlanService {
