+++ Features/Sources/SavingsPlanModule/Screens/Add/SavingPlansAddScreen+ViewModel.swift
@@ -9,6 +9,7 @@ import Foundation
 import SwiftUI
 import CoreModule
 import Dependencies
+import Models
 
 extension SavingPlansAddScreen {
