+++ Features/Sources/SavingsPlanModule/SuccesfullManager+Extensions.swift
@@ -8,6 +8,7 @@
 import Foundation
 import CoreModule
 import SwiftUI
+import Models
 
 public extension SuccessfullModalManager {
